make session persistent even if WHMCS User logged in to another client account
The delimma here, is that in WHMCS 8.x, a User can switch to another Client that they have access to.

This eventually leads to Client A inheriting Client B data in chatwoot, because upon the switch, no SessionReset is being triggered. This eventually leads to mixed up identities of users currently on chat.

The idea is to have the sdk pass persistent data in different case scenarios. This commit assumes that a User has 1 OWNED client at the moment, so it makes the sdk pulls the data of the OWNED client only, even if User is logged in to another client account on which they have access.

The logic should be improved, because a User can theoritically OWN multiple client accounts in WHMCS.