Introduce Version 2.0.0 (#1)
* Updated

* Update chatwoot.php

* Update hooks.php

* Update hooks.php

* Update hooks.php

* Update hooks.php

* Update chatwoot.php

* Update hooks.php

* "echo" for logout code

* Update hooks.php

* Update hooks.php

* Update hooks.php

* Update hooks.php

* Update hooks.php

* Update hooks.php

* Update chatwoot.php

* Update hooks.php

* Update hooks.php

* Update hooks.php

* Update hooks.php

* Update hooks.php

* Update hooks.php

* update

* possible workaround for WHMCS v8 logout obstacle

* ver updt

* Update hooks.php

* make signing_hash a fixed value to avoid accedintal deletion of this string

* add buuble design and lang options

* add a function to convert WHMCS lang name to lang code

* make session persistent even if WHMCS User logged in to another client account

The delimma here, is that in WHMCS 8.x, a User can switch to another Client that they have access to.

This eventually leads to Client A inheriting Client B data in chatwoot, because upon the switch, no SessionReset is being triggered. This eventually leads to mixed up identities of users currently on chat.

The idea is to have the sdk pass persistent data in different case scenarios. This commit assumes that a User has 1 OWNED client at the moment, so it makes the sdk pulls the data of the OWNED client only, even if User is logged in to another client account on which they have access.

The logic should be improved, because a User can theoritically OWN multiple client accounts in WHMCS.

* Update chatwoot.php

* Update chatwoot.php

* Update hooks.php

* Update hooks.php

* Update hooks.php

* Update hooks.php